applying comments from review
